Add prefix and suffix to boolean columns
